// types/activity.ts
export interface TrackingEvent {
  $id: string;
  campaign_id: string;
  recipient_id?: string;
  link_id?: string;
  email: string;
  event_type: "open" | "click" | "sent" | "failed" | "bounce";
  link_url?: string;
  user_agent?: string;
  ip_address?: string;
  user_email: string;
  created_at: string;
  metadata?: string;
}